Hide implementation details in logging module
